+++ src/hapi/pipelines/database/org.py
@@ -6,10 +6,10 @@ from typing import Dict
 from hapi_schema.db_org import DBOrg
 from hdx.scraper.utilities.reader import Read
 from hdx.utilities.dictandlist import dict_of_sets_add
+from hdx.utilities.text import normalise
 from sqlalchemy.orm import Session
 
 from ..utilities.batch_populate import batch_populate
-from ..utilities.mappings import clean_text
 from .base_uploader import BaseUploader
 
 logger = logging.getLogger(__name__)
@@ -57,8 +57,8 @@ class Org(BaseUploader):
         org_type,
     ):
         key = (
-            clean_text(acronym),
-            clean_text(org_name),
+            normalise(acronym),
+            normalise(org_name),
         )
         if key in self.data:
             org_type_old = self.data[key][2]
@@ -68,8 +68,8 @@ class Org(BaseUploader):
             return
         self.data[
             (
-                clean_text(acronym),
-                clean_text(org_name),
+                normalise(acronym),
+                normalise(org_name),
             )
         ] = [acronym, org_name, org_type]
 
@@ -93,9 +93,9 @@ class Org(BaseUploader):
         org_map_info = org_name_map.get(org_name)
         if not org_map_info:
             org_name_map_clean = {
-                clean_text(on): org_name_map[on] for on in org_name_map
+                normalise(on): org_name_map[on] for on in org_name_map
             }
-            org_name_clean = clean_text(org_name)
+            org_name_clean = normalise(org_name)
             org_map_info = org_name_map_clean.get(org_name_clean)
         if not org_map_info:
             return {"#org+name": org_name}