+++ src/hapi/pipelines/utilities/mappings.py
@@ -1,8 +1,7 @@
-from re import sub
 from typing import Dict
 
-from hdx.location.names import clean_name
 from hdx.location.phonetics import Phonetics
+from hdx.utilities.text import normalise
 
 MATCH_THRESHOLD = 5
 
@@ -30,8 +29,8 @@ def get_code_from_name(
     code = code_lookup.get(name)
     if code:
         return code, name, False
-    name_clean = clean_text(name)
-    clean_lookup = {clean_text(c): code_lookup[c] for c in code_lookup}
+    name_clean = normalise(name)
+    clean_lookup = {normalise(c): code_lookup[c] for c in code_lookup}
     code = clean_lookup.get(name_clean)
     if code:
         return code, name_clean, False
@@ -54,10 +53,3 @@ def get_code_from_name(
     name = names[name_index]
     code = code_lookup.get(name, code_mapping.get(name))
     return code, name_clean, True
-
-
-def clean_text(text: str) -> str:
-    text_clean = clean_name(text)
-    text_clean = sub(r"[^'a-zA-Z0-9\s]", " ", text_clean)
-    text_clean = sub(" +", " ", text_clean)
-    return text_clean.strip()